fix HDFS HA can't work on DEBUG mode
The Describe the bug
when the using HDFS HA nameserivce as the uri, the port will be 0.
hdfsBuilderSetNameNodePort will be called to set the port.
hdfsBuilderSetNameNodePort call asset to check if the port is greater
than 0.
So in Release mode, it works OK. In the Debug mode, the asset will fail.

How to reproduce
when compiler the Clickhouse, use DEBUG mode, it will throw error when
using HDFS HA nameservice url