Fix crash in corner case of MergeTreeRangeReader::ReadResult::shrink
It looks like most times the original number of rows per granule is
pretty much consistent across all granules, except for the last one,
however there seems to be cases (Issue #7713) where that's not the case.

The only way I have to reproduce the crash involves using proprietary
data, and unfortunately I haven't been able to come up with a test case
that would highlight the problem - just like the person who opened issue
7713. I can't share my dataset.

Reading the code, it looks pretty obvious to me here - unless I'm
missing something, that the intention here is to copy the corresponding
granules to one another so it makes sense to index
rows_per_granule_original with the same variable as rows_per_granule.

Signed-off-by: Baudouin Giard <bgiard@bloomberg.net>