Switch cppkafka to upstream and bump to upstream/master
All changes in ClickHouse-Extra has been merged into upstream no need in
using the fork.

There are not a lot of changes for now, and cppkafka does not looks like
in the active development stage, but just in case better to sync with
upstream.

FWIW cppkafka usage is pretty minimal and it's usage is questionable,
since librdkafka is under active development while cppkafka is not, and
later still does not support some features that is used in clickhouse:
- interceptors (for logging proxying)