AsynchronousMetricsLog - add improved tests
This adds integration tests to test the asynchronous_metric_log table
for event_time_microseconds field.

Asynchronous metrics are updated once every 60s, so this setting needs
to be overridden for having those metrics available in time so that the
tests can run. So the `asynchronous_metric_update_period_s` setting is
being overriden to be 2s.