Do not cache dictionary for dictGet*/dictHas*
There are places where ExpressionActionsPtr is cached
(StorageDistributed caching it for sharding_key_expr and
optimize_skip_unused_shards), and if the dictionary will be cached
within "query" then cached ExpressionActionsPtr will always have first
version of the query and the dictionary will not be updated after
reload.

For example this will fix dictGet in sharding_key (and similar places,
i.e. when the function context is stored permanently)

Fixes: 01527_dist_sharding_key_dictGet_reload