d/compat: bump compatibility level to 10
Official CI builds uses ubuntu:20.04, which has the following:
- debhelper: 12.10ubuntu1
- dpkg-dev: 1.19.7ubuntu3 (should not be relevant, but let's keep it)

While debian:testing has:
- debhelper: 13.3
- dpkg: 1.20.5 (should not be relevant, but let's keep it)

In debhelper 13 compat level <= 9 has been deprecated [1], hence the
warning:

    warning: Compatibility levels before 10 are deprecated (level 8 in use)

  [1]: https://salsa.debian.org/debian/debhelper/-/commit/427594c41330917401356ebed2a951e0e9143256

And this is not only warning but also compat level implicitly will be
changed to 10. For now this is only for debian:testing but in future it
will be relevant for newer ubuntu, and hence official builds.